refactor: extract shared CLI arg parsing
Move readRepeatedFlag/splitList/normalizeExperimentName/readSuiteFilters into apps/framework/lib/cli-args.ts, shared by run-eval and export-results.
